Update README.md with like a million changes idk
Added a description to all commands in the Homes section.
Added a description to all commands in the Teleportation section.
Added a description to all commands in the Marriage section.
Moved Graves to Gameplay Changes section.
Added a description to all features in the Grave section.
Added (to do) in Loot regeneration and Quick Inventory Dumping in the Misc section (Planning to give a more in detail description of such)
Added Mail System (to do) and Interactive Chat (to do) in the Misc section
Added another sentence to `/togglepvp`

Most information is taken from the `/help` command in the Minecraft Server, alongside my own memory. Some info might be incorrect/missing, no guarantees!